Contextualizacao das buscas

+++ src/main/java/com/sistemaAutomotivo/SistemaAutomotivo/modulos/equipes/services/EquipeServiceImpl.java
@@ -3,7 +3,6 @@ package com.sistemaAutomotivo.SistemaAutomotivo.modulos.equipes.services;
 import java.util.ArrayList;
 import java.util.List;
 
-import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -11,7 +10,7 @@ import com.sistemaAutomotivo.SistemaAutomotivo.modulos.equipes.dto.EquipeDTO;
 import com.sistemaAutomotivo.SistemaAutomotivo.modulos.equipes.entities.Equipe;
 import com.sistemaAutomotivo.SistemaAutomotivo.modulos.equipes.repositories.EquipeRepository;
 import com.sistemaAutomotivo.SistemaAutomotivo.modulos.funcionarios.entities.Funcionario;
-import com.sistemaAutomotivo.SistemaAutomotivo.modulos.funcionarios.repositories.FuncionarioRepository;
+import com.sistemaAutomotivo.SistemaAutomotivo.modulos.funcionarios.services.FuncionarioService;
 
 @Service
 public class EquipeServiceImpl implements EquipeService {
@@ -20,53 +19,105 @@ public class EquipeServiceImpl implements EquipeService {
     private EquipeRepository equipeRepository;
 
     @Autowired
-    private FuncionarioRepository funcionarioRepository;
+    private FuncionarioService funcionarioService;
 
     // CREATE
     @Override
-    public Equipe saveEquipe(EquipeDTO equipe) {
-        return equipeRepository.save(DTOtoEquipe(equipe));
+    public Equipe saveEquipe(EquipeDTO equipeDTO) {
+
+        /* verifica se o nome da equipe já existe */
+        if (equipeRepository.existsByNome(equipeDTO.nome())) {
+            throw new RuntimeException("Nome de equipe já cadastrado!");
+        }
+
+        Equipe novaEquipe = DTOtoEquipe(equipeDTO);
+
+        /* pega o funcionario responsavel e adiciona a equipe */
+        Funcionario funcionarioResponsavel = funcionarioService.findByCpf(equipeDTO.cpfResponsavel());
+        novaEquipe.setResponsavel(funcionarioResponsavel);
+        novaEquipe.getMembros().add(funcionarioResponsavel);
+
+        return equipeRepository.save(novaEquipe);
     }
 
     // READ
     @Override
     public Equipe findById(Integer idEquipe) {
+
+        if (!(equipeRepository.existsById(idEquipe))) {
+            throw new RuntimeException("Nenhuma equipe com esse id encontrada!");
+        }
+
         return equipeRepository.findById(idEquipe).get();
     }
 
     @Override
     public List<Equipe> findAllEquipes() {
+
+        if (equipeRepository.findAll().isEmpty()) {
+            throw new RuntimeException("Nenhuma equipe encontrada!");
+        }
+
         return equipeRepository.findAll();
     }
 
     @Override
-    public List<Funcionario> findAllMembros(Integer idEquipe) {
-        Equipe equipe = equipeRepository.findById(idEquipe).get();
+    public List<Funcionario> findAllMembros(String nome) {
+        Equipe equipe = findByNome(nome);
 
         return equipe.getMembros();
     }
 
+    @Override
+    public Equipe findByNome(String nome) {
+
+        nome = nome.replace("_", " ");
+
+        return equipeRepository.findByNome(nome)
+                .orElseThrow(() -> new RuntimeException("Nenhuma equipe com esse nome foi encontrada!"));
+    }
+
     // UPDATE
     @Override
-    public Equipe updateEquipeById(Integer idEquipe, EquipeDTO equipeDTO) {
-        // pega a equipe existente com base no id
-        Equipe equipeExistente = findById(idEquipe);
+    public Equipe updateEquipeByNome(String nome, EquipeDTO equipeDTO) {
+        // pega a equipe existente com base no nome
+        Equipe equipeExistente = findByNome(nome);
 
         // constroi a equipe atualizada
-        Equipe equipeAtualizada = DTOtoEquipe(equipeDTO);
-        equipeAtualizada.setId_equipe(equipeExistente.getId_equipe());
+        /*
+         * Equipe equipeAtualizada = DTOtoEquipeUpdate(equipeDTO);
+         * equipeAtualizada.setId_equipe(equipeExistente.getId_equipe());
+         */
+
+        // atualiza nome e setor
+        equipeExistente.setNome(equipeDTO.nome());
+        equipeExistente.setSetor(equipeDTO.setor());
+
+        // caso seja definido um novo responsavel
+        if (equipeExistente.getResponsavel().getCpf() != equipeDTO.cpfResponsavel()) {
+            Funcionario novoResponsavel = funcionarioService.findByCpf(equipeDTO.cpfResponsavel());
+            equipeExistente.setResponsavel(novoResponsavel);
+
+            if (!(equipeExistente.getMembros().contains(novoResponsavel))) {
+                equipeExistente.getMembros().add(novoResponsavel);
+            }
+        }
 
         // copia as propriedades da equipe atualizada para a existente
-        BeanUtils.copyProperties(equipeAtualizada, equipeExistente);
+        // BeanUtils.copyProperties(equipeAtualizada, equipeExistente);
 
         // salva e retorna novamente a equipe existente
         return equipeRepository.save(equipeExistente);
     }
 
     @Override
-    public Funcionario adicionarFuncionario(Integer idEquipe, Integer idFuncionario) {
-        Equipe equipeExistente = equipeRepository.findById(idEquipe).get();
-        Funcionario funcionario = funcionarioRepository.findById(idFuncionario).get();
+    public Funcionario adicionarFuncionario(String nome, String cpfFuncionario) {
+        Equipe equipeExistente = findByNome(nome);
+        Funcionario funcionario = funcionarioService.findByCpf(cpfFuncionario);
+
+        if (equipeExistente.getMembros().contains(funcionario)) {
+            throw new RuntimeException("Funcionario já é parte da equipe!");
+        }
 
         equipeExistente.getMembros().add(funcionario);
         equipeRepository.save(equipeExistente);
@@ -75,9 +126,13 @@ public class EquipeServiceImpl implements EquipeService {
     }
 
     @Override
-    public Funcionario removerFuncionario(Integer idEquipe, Integer idFuncionario) {
-        Equipe equipeExistente = equipeRepository.findById(idEquipe).get();
-        Funcionario funcionario = funcionarioRepository.findById(idFuncionario).get();
+    public Funcionario removerFuncionario(String nome, String cpfFuncionario) {
+        Equipe equipeExistente = findByNome(nome);
+        Funcionario funcionario = funcionarioService.findByCpf(cpfFuncionario);
+
+        if (equipeExistente.getResponsavel().getCpf() == funcionario.getCpf()) {
+            throw new RuntimeException("O líder da equipe não pode ser excluído!");
+        }
 
         equipeExistente.getMembros().remove(funcionario);
 
@@ -88,8 +143,13 @@ public class EquipeServiceImpl implements EquipeService {
 
     // DELETE
     @Override
-    public Equipe deleteEquipeById(Integer idEquipe) {
-        Equipe equipe = equipeRepository.findById(idEquipe).get();
+    public Equipe deleteEquipeByNome(String nome) {
+        Equipe equipe = findByNome(nome);
+
+        if (!(equipe.getServicos_orcamentos().isEmpty())) {
+            throw new RuntimeException("Equipe não pode ser excluída caso seja responsavel por um Servico_Orcamento!");
+        }
+
         equipeRepository.delete(equipe);
         return equipe;
     }
@@ -97,22 +157,18 @@ public class EquipeServiceImpl implements EquipeService {
     // metodos auxiliares
     public Equipe DTOtoEquipe(EquipeDTO equipeDTO) {
 
-        Funcionario funcionario = funcionarioRepository
-                .findById(equipeDTO.idResponsavel())
-                .get();
-
-        List<Funcionario> membros = new ArrayList<>();
-        membros.add(funcionario);
-
         Equipe equipe = new Equipe(null,
                 equipeDTO.nome(),
                 equipeDTO.setor(),
-                funcionario,
-                membros,
+                null,
+                new ArrayList<>(),
                 new ArrayList<>());
 
         return equipe;
     }
 
-    
+    public Equipe DTOtoEquipeUpdate(EquipeDTO equipeDTO) {
+        return null;
+    }
+
 }
