+++ src/main/java/com/sistemaAutomotivo/SistemaAutomotivo/modulos/funcionarios/services/FuncionarioServiceImpl.java
@@ -3,7 +3,6 @@ package com.sistemaAutomotivo.SistemaAutomotivo.modulos.funcionarios.services;
 import java.util.ArrayList;
 import java.util.List;
 
-import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -14,57 +13,90 @@ import com.sistemaAutomotivo.SistemaAutomotivo.modulos.funcionarios.repositories
 
 @Service
 public class FuncionarioServiceImpl implements FuncionarioService {
-    
+
     @Autowired
     private FuncionarioRepository funcionarioRepository;
-    
+
     // CREATE
     @Override
     public Funcionario saveFuncionario(FuncionarioDTO funcionarioDTO) {
-        // colocar if existsByCPF
+
+        if (funcionarioRepository.existsByCpf(funcionarioDTO.cpf())) {
+            throw new RuntimeException("CPF de funcionário já cadastrado!");
+        }
+        if (funcionarioRepository.existsByEmail(funcionarioDTO.email())) {
+            throw new RuntimeException("Email de funcionário já cadastrado!");
+        }
+
         return funcionarioRepository.save(DTOtoFuncionario(funcionarioDTO));
     }
 
     // READ
     @Override
     public List<Funcionario> findAllFuncionarios() {
-       // colocar if isEmpty
+
+        if (funcionarioRepository.findAll().isEmpty()) {
+            throw new RuntimeException("Nenhum funcionário encontrado");
+        }
+
         return funcionarioRepository.findAll();
     }
 
     @Override
     public Funcionario findById(Integer id) {
+
+        if (!(funcionarioRepository.existsById(id))) {
+            throw new RuntimeException("Nenhum funcionario encontrado!");
+        }
+
         return funcionarioRepository.findById(id).get();
     }
 
     @Override
-    public List<Equipe> findAllEquipes(Integer idFuncionario) {
-        Funcionario funcionario = funcionarioRepository.findById(idFuncionario).get();
+    public List<Equipe> findAllEquipes(String cpf) {
+        Funcionario funcionario = findByCpf(cpf);
+
+        if (funcionario.getEquipes().isEmpty()) {
+            throw new RuntimeException("Funcionário não está em nenhuma equipe!");
+        }
 
         return funcionario.getEquipes();
     }
 
+    @Override
+    public Funcionario findByCpf(String cpf) {
+        return funcionarioRepository.findByCpf(cpf)
+                .orElseThrow(() -> new RuntimeException("Nenhum funcionario encontrado!"));
+    }
+
     // UPDATE
     @Override
-    public Funcionario updateById(Integer id, FuncionarioDTO funcionarioDTO) {
-        // pega o funcionario existente com base no id
-        Funcionario funcionarioExistente = findById(id);
-        
+    public Funcionario updateByCpf(String cpf, FuncionarioDTO funcionarioDTO) {
+        // pega o funcionario existente com base no cpf
+        Funcionario funcionarioExistente = findByCpf(cpf);
+
         // constroi o funcionario atualizado
-        Funcionario funcionarioAtualizado = DTOtoFuncionario(funcionarioDTO);
-        funcionarioAtualizado.setId_funcionario(funcionarioExistente.getId_funcionario());
+        funcionarioExistente.setNome(funcionarioDTO.nome());
+        funcionarioExistente.setApelido(funcionarioDTO.apelido());
+        funcionarioExistente.setCpf(funcionarioDTO.cpf());
+        funcionarioExistente.setEndereco(funcionarioDTO.endereco());
+        funcionarioExistente.setOficio(funcionarioDTO.oficio());
+        funcionarioExistente.setCargaHorariaSemanal(funcionarioDTO.cargaHorariaSemanal());
+        funcionarioExistente.setEmail(funcionarioDTO.email());
+        funcionarioExistente.setTelefone(funcionarioDTO.telefone());
 
-        // copia as propriedades do funcionario atualizado para o existente
-        BeanUtils.copyProperties(funcionarioAtualizado, funcionarioExistente);
-
-        //salva e retorna novamente o funcionario existente
+        // salva e retorna novamente o funcionario existente
         return funcionarioRepository.save(funcionarioExistente);
     }
 
     // DELETE
     @Override
-    public Funcionario deleteById(Integer id) {
-        Funcionario funcionarioExcluido = funcionarioRepository.findById(id).get();
+    public Funcionario deleteByCpf(String cpf) {
+        Funcionario funcionarioExcluido = findByCpf(cpf);
+
+        if (!(funcionarioExcluido.getEquipes().isEmpty())) {
+            throw new RuntimeException("Funcionario não pode ser excluído caso faça parte de alguma equipe!");
+        }
 
         funcionarioRepository.delete(funcionarioExcluido);
 
@@ -74,18 +106,18 @@ public class FuncionarioServiceImpl implements FuncionarioService {
     // metodos auxiliares
     Funcionario DTOtoFuncionario(FuncionarioDTO funcionarioDTO) {
 
-        // o funcionario eh instanciado sem fazer de nenhuma equipe
+        // o funcionario eh instanciado sem fazer parte de nenhuma equipe
         Funcionario funcionario = new Funcionario(
-            null,
-            funcionarioDTO.nome(),
-            funcionarioDTO.apelido(),
-            funcionarioDTO.cpf(),
-            funcionarioDTO.endereco(),
-            funcionarioDTO.oficio(),
-            funcionarioDTO.cargaHorariaSemanal(),
-            funcionarioDTO.email(),
-            funcionarioDTO.telefone(),
-            new ArrayList<>());
+                null,
+                funcionarioDTO.nome(),
+                funcionarioDTO.apelido(),
+                funcionarioDTO.cpf(),
+                funcionarioDTO.endereco(),
+                funcionarioDTO.oficio(),
+                funcionarioDTO.cargaHorariaSemanal(),
+                funcionarioDTO.email(),
+                funcionarioDTO.telefone(),
+                new ArrayList<>());
 
         return funcionario;
     }
